refactor: simplify logo URL fallback parsing
Agent-Logs-Url: https://github.com/SamWalts/theflightwall_oss_forked/sessions/acd360a5-468c-46ec-b4e5-077d185d0359

+++ firmware/adapters/NeoMatrixDisplay.cpp
@@ -131,7 +131,7 @@ String NeoMatrixDisplay::makeAirlineLogoBadgeText(const FlightInfo &f)
 
     if (badge.length() == 0)
     {
-        for (size_t i = 0; i < (size_t)f.airline_display_name_full.length() && badge.length() < 3; ++i)
+        for (size_t i = 0; i < f.airline_display_name_full.length() && badge.length() < 3; ++i)
         {
             char c = f.airline_display_name_full[i];
             if (isalpha((unsigned char)c))
